Restructured searchbar export and file to a .jsx | Implemented styled-components in SearchBar.jsx
