--- src/components/common/searchbar/searchbarStyles.css
@@ -1,31 +1,0 @@
-.nope{
-    display: none;
-}
-
-.searchOuter{
-    background-color: white;
-    padding: 20px;
-    display: flex;
-    justify-content: space-between;
-    border: 1px solid lightgray;
-    padding-bottom: 10px
-}
-.searchBar{
-    padding: 0 5px 0 50px;
-    max-width: 300px;
-}
-
-.add-item-button{
-    margin-right: 20px;
-    background-color: rgb(19, 75, 29);
-    color: white;
-}
-.searchBtns {
-    display: flex;
-    background-color: white;
-}
-
-.searchBtns button{
-    background-color: rebeccapurple;
-    color: white;
-}
